WinSetup: moved WinSetup helper functions to winapihelpers.cpp
