+++ Solutions/Engine.App/Engine.App.vcxproj
@@ -652,6 +652,7 @@
     <ClCompile Include="..\..\Engine\platform\base\agsplatformdriver.cpp" />
     <ClCompile Include="..\..\Engine\platform\base\sys_main.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\acplwin.cpp" />
+    <ClCompile Include="..\..\Engine\platform\windows\setup\winapihelpers.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\win_ex_handling.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\debug\namedpipesagsdebugger.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\gfx\ali3dd3d.cpp" />
@@ -921,6 +922,7 @@
     <ClInclude Include="..\..\Engine\platform\base\agsplatformdriver.h" />
     <ClInclude Include="..\..\Engine\platform\base\sys_main.h" />
     <ClInclude Include="..\..\Engine\platform\windows\debug\namedpipesagsdebugger.h" />
+    <ClInclude Include="..\..\Engine\platform\windows\setup\winapihelpers.h" />
     <ClInclude Include="..\..\Engine\platform\windows\win_ex_handling.h" />
     <ClInclude Include="..\..\Engine\platform\windows\gfx\ali3dd3d.h" />
     <ClInclude Include="..\..\Engine\platform\windows\setup\winsetup.h" />
