+++ Solutions/Engine.App/Engine.App.vcxproj.filters
@@ -810,6 +810,9 @@
     <ClCompile Include="..\..\Engine\media\video\theora_player.cpp">
       <Filter>Source Files\media\video</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Engine\platform\windows\setup\winapihelpers.cpp">
+      <Filter>Source Files\setup</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Engine\ac\asset_helper.h">
@@ -1630,6 +1633,9 @@
     </ClInclude>
     <ClInclude Include="..\..\Engine\media\video\theora_player.h">
       <Filter>Header Files\media\video</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Engine\platform\windows\setup\winapihelpers.h">
+      <Filter>Header Files\setup</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
